fix(etl): paginate session-text rebuild in extract_block_context
extract_block_context still called the unbounded get_frame_full_texts +
build_session_text over the current run's block range. On a first-run
backfill after long downtime, a single batch can span many frames for
one long-lived app block, spiking RSS the same way the whole-session
rebuild path did before PR #566.

Moves rebuild_session_text_paginated into text_merge.rs (shared by both
the whole-session close path in block_ops.rs and this per-run extraction
path) so extract_block_context reads frames in bounded pages instead of
materializing the whole range at once. Output is unchanged — same
parity test (test_session_text_paginated_rebuild_matches_non_paginated_fetch)
already covers this code path since extractor.rs now calls the identical
paginated function.
